Progress Update 2
Added tons of documentation comments in the GameManager script, stuff that I am working on next.
Also added in the GlassJar prefab from a free models asset that holds the coins from the UpdateMoney() method
The problem is that I had to add box colliders surrounding the jar so it doesn't really exactly hold the coins inside the glass texture perfectly because I needed there to be primitive colliders to collide with the primitive collider of the coin, the mesh collider was simply not working with the coin.